+++ CBench/metrics/relativeError.hpp
@@ -109,7 +109,7 @@ inline void relativeError::execute(void *original, void *approx, size_t n) {
 		if (total_max_rel_err != 0)
 		{
 			std::vector<float>histogram;
-			int numBins = 1024;
+			size_t numBins = 1024;
 			std::vector<int> localHistogram(numBins,0);
 			double binSize = total_max_rel_err / numBins;
